+++ src/dojo/sql/budgeting/insert_cash_account_detail.sql
@@ -1,10 +1,12 @@
 INSERT INTO cash_account_details (
     detail_id,
-    account_id
+    account_id,
+    interest_rate_apy
 )
 SELECT
     $detail_id,
-    $account_id
+    $account_id,
+    $interest_rate_apy
 WHERE NOT EXISTS (
     SELECT 1 FROM cash_account_details
     WHERE