+++ src/dojo/sql/budgeting/insert_credit_account_detail.sql
@@ -1,10 +1,16 @@
 INSERT INTO credit_account_details (
     detail_id,
-    account_id
+    account_id,
+    apr,
+    card_type,
+    cash_advance_apr
 )
 SELECT
     $detail_id,
-    $account_id
+    $account_id,
+    $apr,
+    $card_type,
+    $cash_advance_apr
 WHERE NOT EXISTS (
     SELECT 1 FROM credit_account_details
     WHERE