+++ src/dojo/sql/budgeting/insert_investment_account_detail.sql
@@ -1,10 +1,16 @@
 INSERT INTO investment_account_details (
     detail_id,
-    account_id
+    account_id,
+    risk_free_sweep_rate,
+    is_self_directed,
+    tax_classification
 )
 SELECT
     $detail_id,
-    $account_id
+    $account_id,
+    $risk_free_sweep_rate,
+    $is_self_directed,
+    $tax_classification
 WHERE NOT EXISTS (
     SELECT 1 FROM investment_account_details
     WHERE